+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Changed
+- updated unit tests to PHPUnit 6.
+
 ## 0.2.1 - 2017-12-28
 
 ### Fixed