+++ CHANGELOG.md
@@ -6,8 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## 0.2.2 - 2018-02-01
+
 ### Changed
 - updated unit tests to PHPUnit 6.
+
+### Fixed
+- setting timezone on immutable instance created from UTC post time.
 
 ## 0.2.1 - 2017-12-28